Don't add every core/ subdir to -I path.
We should have the Arduino core itself on the include search path
but not every subdir thereof.

Requires copying the Arduino core headers into the build/core/ dir,
and build/core/ must be at the head of the search path so that when
building the core, `#include<Arduino.h>` and `#include "Arduino.h"`
are always resolved to the same (local) file rather than causing
redefinition errors.