Use explicit profiles rather than arduino-cli for build cmds.
The old system of trying to scrape stdout of arduino-cli no longer
works in v1.x. That's ok; we only have so many board vendors and the
parameters can just be looked up. Use arduino-cli to determine the
connected board / fqbn, then resolve that to a "profile" of more
properties to include in the build.